backend/python-ai/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai-whisper==20231117
openai==1.10.0
pydantic==2.5.3
